+++ controller/admin_controller.php
@@ -258,11 +258,13 @@ class admin_controller
 			'ACP_IMAGEUPLOAD_EXT'					=> $imageupload_allowed_extensions,
 			'ACP_IMAGEUPLOAD_INDEX_ENABLE'			=> $this->config['imageupload_index_enable'],
 			'ACP_IMAGEUPLOAD_NUMBER'				=> $this->config['imageupload_number'],
+			'ACP_IMAGEUPLOAD_MAX_NUMBER'			=> $max_filesize,
 			'ACP_IMAGEUPLOAD_ALLOWED_SIZE'			=> $this->user->lang('ACP_IMAGEUPLOAD_NEW_DOWNLOAD_SIZE', $max_filesize, $unit),
 			'ACP_TOTAL_IMAGES'						=> $this->user->lang('ACP_MULTI_IMAGES', (int) $total_imageupload),
 			'ACP_IMAGEUPLOAD_CHAT_ENABLE'			=> $this->config['imageupload_chat_enable'],
 			'ACP_IMAGEUPLOAD_POST_ENABLE'			=> $this->config['imageupload_post_enable'],
 			'ACP_IMAGEUPLOAD_POSTTAB_ENABLE'		=> $this->config['imageupload_posttab_enable'],
+			'ACP_IMAGEUPLOAD_MULTI_ENABLE'			=> $this->config['imageupload_multiupload_enable'],
 			'ACP_IMAGEUPLOAD_ENABLE_DIRECT_LINK'	=> $this->config['imageupload_enable_direct_link'],
 			'ACP_IMAGEUPLOAD_ENABLE_URL_LINK'		=> $this->config['imageupload_enable_url_link'],
 			'ACP_IMAGEUPLOAD_ENABLE_IMG_LINK'		=> $this->config['imageupload_enable_img_link'],
@@ -348,6 +350,7 @@ class admin_controller
 		$this->config->set('imageupload_chat_enable', $this->request->variable('imageupload_chat_enable', 0));
 		$this->config->set('imageupload_post_enable', $this->request->variable('imageupload_post_enable', 0));
 		$this->config->set('imageupload_posttab_enable', $this->request->variable('imageupload_posttab_enable', 0));
+		$this->config->set('imageupload_multiupload_enable', $this->request->variable('imageupload_multiupload_enable', 0));
 		$this->config->set('imageupload_enable_direct_link', $this->request->variable('imageupload_enable_direct_link', 0));
 		$this->config->set('imageupload_enable_url_link', $this->request->variable('imageupload_enable_url_link', 0));
 		$this->config->set('imageupload_enable_img_link', $this->request->variable('imageupload_enable_img_link', 0));