+++ language/en/permissions_imageupload.php
@@ -33,8 +33,9 @@ if (empty($lang) || !is_array($lang))
 // ’ » “ ” …
 
 $lang = array_merge($lang, [
-	'ACL_U_IMAGE_UPLOAD'		=> 'Can upload images in Image Upload',
 	'ACL_U_IMAGEUPLOAD'			=> 'Image Upload',
+	'ACL_U_IMAGE_UPLOAD'		=> 'Can upload images',
 	'ACL_U_IMAGE_DELETE'		=> 'Can delete images in UCP',
 	'ACL_U_IMAGE_UPLOAD_UCP'	=> 'Can see images in UCP profile tab',
+	'ACL_U_IMAGE_UPLOAD_MULTI'	=> 'Can upload multiple images',
 ]);